requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
aiohttp==3.9.3
